+++ packages/schema/index.ts
@@ -1,4 +1,5 @@
 import { makeSchema } from "@nexus/schema";
+import { nexusPrisma } from "nexus-plugin-prisma";
 import { resolve } from "path";
 import { config } from "@packages/config";
 import * as types from "./types";
@@ -8,6 +9,15 @@ const shouldGenerateArtifacts =
 
 export const schema = makeSchema({
   types,
+  plugins: [
+    nexusPrisma({
+      experimentalCRUD: true,
+      outputs: {
+        typegen: resolve(__dirname, "generated/nexus-prisma-types.ts"),
+      },
+      shouldGenerateArtifacts,
+    }),
+  ],
   outputs: {
     schema: resolve(__dirname, "generated/nexus-schema.graphql"),
     typegen: resolve(__dirname, "generated/nexus-types.ts"),